requirements.txt
streamlit>=1.37.0
pandas>=2.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
numpy>=1.24.0
scikit-learn>=1.3.0
sentence-transformers>=2.7.0
arxiv>=2.1.0
requests>=2.31.0
python-dotenv>=1.0.0
groq>=0.9.0
PyMuPDF>=1.24.0
rank-bm25>=0.2.2
